Fix for GHC 6.10 compatibility
From: Neil Mitchell
Date: Mon, 8 Dec 2008

Hi John,

I'm using HsLogger, but unfortunately it fails to build with GHC 6.10.
You can see the error log here:

http://hackage.haskell.org/packages/archive/hslogger/1.0.6/logs/failure/ghc-6.10

The problem is the use of the new exceptions stuff, which has
generalised the type of exceptions, giving ambiguous types when you
try and show it. The simplest solution is to change the line in the
.cabal file to:

        build-depends: base >= 3 && < 4, containers, directory, process

i.e. use the GHC 6.8 compatible base even on GHC 6.10. This should
work for both versions.

Thanks

Neil