fixed encoding problem for stream handlers
Starting with GHC 6.12.1, hPutStrLn (as used by the stream handler in
src/System/Log/Handler/Simple.hs) encodes a given String according to the
platform locale. If encoding fails, you get an exception. The patch adds
code for catching any exception thrown while writing the log message to a
handle. If such an exception is thrown, an appropriate error message is
written to the log. The error message is guaranteed to be encodable in
ASCII.

Conflicts:

	src/System/Log/Handler/Simple.hs