Simple handler: evaluate message before taking lock
Because of laziness, the `msg` string in `streamHandler` can contain
arbitrary computations. In the previous state of affairs, the handler
would first take the lock then `putStrLn` the message. So all
the (pure) computation would happen in the critical section, and it
couldn't guarantee timely release of the lock.

Calling `deepseq` before the lock is taken makes sure that the locked
section does only what it must do: print the string.